#8 - change same react version - rename file customer/app to customer/CustomerListView - format using react-intl
