+++ package.json
@@ -42,7 +42,7 @@
     ]
   },
   "devDependencies": {
-    "@types/react": "^17.0.3",
+    "@types/react": "^17.0.2",
     "@types/react-dom": "^17.0.2",
     "@types/react-router-dom": "^5.1.7",
     "typescript": "^4.2.3"
